laptop/requirements.txt
pygame
PyOpenGL
numpy
zeroconf
